fix another test
