+++ jupyter_releaser/tests/conftest.py
@@ -3,9 +3,11 @@
 import json
 import os
 import os.path as osp
+import tempfile
 from pathlib import Path
 
 from click.testing import CliRunner
+from ghapi.core import GhApi
 from pytest import fixture
 
 from jupyter_releaser import cli, util
@@ -195,3 +197,28 @@ def mock_github():
     if proc:
         proc.kill()
         proc.wait()
+
+
+@fixture
+def draft_release(mock_github):
+    gh = GhApi(owner="foo", repo="bar")
+    data = dict(
+        version_spec="foo",
+        branch="bar",
+        repo="fizz",
+        since="buzz",
+        since_last_stable=False,
+        version="1.0.0",
+        post_version_spec="dev",
+        post_version_message="hi",
+    )
+    with tempfile.TemporaryDirectory() as d:
+        metadata_path = Path(d) / "metadata.json"
+        with open(metadata_path, "w") as fid:
+            json.dump(data, fid)
+
+        release = gh.create_release(
+            f"v1.0.0", "bar", f"v1.0.0", "hi", True, True, files=[metadata_path]
+        )
+    yield release.html_url
+    gh.repos.delete_release(release.id)
